Quick fix + NovelLibrary
added NovelLibrary
